Add support for building the Windows ffmpeg configuration on Linux!
This adds cross-compiling support for Windows ffmpeg configurations;
it also updates the configurations with the generated results!

This uses a couple of existing Chromium scripts for setting up the
correct MSVC paths and passing the right cflags and ldflags to
chromium such that compilation can succeed.

This also pulls in an upstream patch for configure to make it detect
the proper command line flag type '/' vs '-':
https://patchwork.ffmpeg.org/patch/7469/

There's one further change to the makedef line in configure to
ensure that it's run using the substituted AR and NM commands
instead of the defaults lib and dumpbin.

BUG=783021
TEST=chrome works!